requirements.txt
aioschedule==0.5.2
anyio==4.0.0
APScheduler==3.10.4
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
pyaes==1.6.1
pyasn1==0.5.1
pymongo==4.6.0
python-decouple==3.8
python-telegram-bot==20.6
pytz==2023.3.post1
rsa==4.9
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
Telethon==1.33.0
tzlocal==5.2
Werkzeug==3.0.1
zipp==3.17.0
